+++ src/Accounts/Accounts/Account/ConnectAzureRmAccount.cs
@@ -133,6 +133,11 @@ namespace Microsoft.Azure.Commands.Profile
         public string GraphAccessToken { get; set; }
 
         [Parameter(ParameterSetName = AccessTokenParameterSet,
+                   Mandatory = false, HelpMessage = "Access token to Microsoft Graph")]
+        [ValidateNotNullOrEmpty]
+        public string MicrosoftGraphAccessToken { get; set; }
+
+        [Parameter(ParameterSetName = AccessTokenParameterSet,
                    Mandatory = false, HelpMessage = "AccessToken for KeyVault Service")]
         [ValidateNotNullOrEmpty]
         public string KeyVaultAccessToken { get; set; }
@@ -338,6 +343,7 @@ namespace Microsoft.Azure.Commands.Profile
                     azureAccount.SetProperty(AzureAccount.Property.AccessToken, AccessToken);
                     azureAccount.SetProperty(AzureAccount.Property.GraphAccessToken, GraphAccessToken);
                     azureAccount.SetProperty(AzureAccount.Property.KeyVaultAccessToken, KeyVaultAccessToken);
+                    azureAccount.SetProperty(Constants.MicrosoftGraphAccessToken, MicrosoftGraphAccessToken);
                     break;
                 case ServicePrincipalCertificateParameterSet:
                 case ServicePrincipalCertificateFileParameterSet: